+++ src/components/text-editor/prosemirror-adapter/prosemirror-adapter.tsx
@@ -363,7 +363,10 @@ export class ProsemirrorAdapter {
             <div class="toolbar">
                 <limel-action-bar
                     ref={(el) => (this.actionBarElement = el)}
-                    accessibleLabel="Toolbar"
+                    accessibleLabel={translate.get(
+                        'editor-menu.toolbar',
+                        this.language
+                    )}
                     actions={this.actionBarItems}
                     onItemSelected={this.handleActionBarItem}
                 />
@@ -385,6 +388,7 @@ export class ProsemirrorAdapter {
                 anchor={this.actionBarElement}
             >
                 <limel-text-editor-link-menu
+                    language={this.language}
                     link={this.link}
                     isOpen={this.isLinkMenuOpen}
                     onLinkChange={this.handleLinkChange}